test: refactor test suite into modular units and add comprehensive effect validation tests.
