+++ src/presets/preset_manager_impl.h
