+++ tests/integration/test_clipboard_preset.cpp
