+++ tests/integration/test_json_serialization.cpp
