+++ tests/integration/test_midi_manager.cpp
