+++ tests/integration/test_recorder.cpp
