+++ tests/integration/test_snapshot_manager.cpp
