+++ tests/test_framework.h
@@ -15,6 +15,13 @@ struct TestResult {
     std::string message;
 };
 
+class Test {
+public:
+    virtual ~Test() = default;
+    virtual void SetUp() {}
+    virtual void TearDown() {}
+};
+
 class TestSuite {
 public:
     static TestSuite& instance() {
@@ -92,6 +99,29 @@ private:
         Register_##name() { TestFramework::TestSuite::instance().add_test(#name, test_##name); } \
     } reg_##name; } \
     static void test_##name()
+
+#define TEST_F(FixtureName, TestName) \
+    class FixtureName##_##TestName : public FixtureName { \
+    public: \
+        void RunTest(); \
+    }; \
+    static void run_fixture_test_##FixtureName##_##TestName() { \
+        FixtureName##_##TestName t; \
+        t.SetUp(); \
+        try { \
+            t.RunTest(); \
+        } catch (...) { \
+            t.TearDown(); \
+            throw; \
+        } \
+        t.TearDown(); \
+    } \
+    namespace { struct Register_##FixtureName##_##TestName { \
+        Register_##FixtureName##_##TestName() { \
+            TestFramework::TestSuite::instance().add_test(#FixtureName "_" #TestName, run_fixture_test_##FixtureName##_##TestName); \
+        } \
+    } reg_##FixtureName##_##TestName; } \
+    void FixtureName##_##TestName::RunTest()
 
 #define ASSERT_TRUE(expr) \
     do { if (!(expr)) { \
