+++ tests/ui/test_gui_manager.cpp
