+++ tests/unit/test_command_history.cpp
